fix warnings on Mac in src/pdf
Fix these class of warnings:
- unused functions
- unused locals
- sign mismatch
- missing function prototypes
- missing newline at end of file
- 64 to 32 bit truncation

The changes prefer to link in dead code in the debug build
with 'if (false)' than to comment it out, but trivial cases
are commented out or sometimes deleted if it appears to be
a copy/paste error.
Review URL: https://codereview.appspot.com/6299047

git-svn-id: http://skia.googlecode.com/svn/trunk@4178 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/pdf/SkPDFPage.cpp
@@ -37,7 +37,7 @@ void SkPDFPage::finalizePage(SkPDFCatalog* catalog, bool firstPage,
 
 off_t SkPDFPage::getPageSize(SkPDFCatalog* catalog, off_t fileOffset) {
     SkASSERT(fContentStream.get() != NULL);
-    catalog->setFileOffset(fContentStream.get(), fileOffset);
+    catalog->setFileOffset(fContentStream.get(), (size_t) fileOffset);
     return fContentStream->getOutputSize(catalog, true);
 }
